Add "tunnel 6T04" to all panel languages
Added translation to all other languages in the panel. And I fixed the flaw in the Russian language a little bit (once again)

P.S. correct me if the translation is crooked or incorrect.

And there is a small question: "Is it possible to leave comments or notes in the translation files and, in general, is it possible to completely rewrite and make it... is it more convenient? or do you not like this idea? Give a feedback pls"